+++ automation/terraform/modules/kubernetes/testnet/kubernetes.tf
@@ -1,8 +1,8 @@
 data "google_client_config" "current" {}
 
 provider "kubernetes" {
-  alias = "testnet_deploy"
-  config_context  = var.k8s_context
+  alias          = "testnet_deploy"
+  config_context = var.k8s_context
 }
 
 resource "kubernetes_namespace" "testnet_namespace" {