+++ automation/terraform/modules/o1-integration/testnet.tf
@@ -24,19 +24,7 @@ module "kubernetes_testnet" {
 
   seed_zone   = "us-west1-a"
   seed_region = "us-west1"
-  seed_configs = [
-    {
-      name               = "seed",
-      class              = "seed",
-      libp2p_secret      = local.seed_peer.secret,
-      external_port      = 10401,
-      node_port          = null,
-      external_ip        = null,
-      private_key_secret = null
-    }
-  ]
-
-  archive_node_count  = 0
+  seed_configs = [local.seed_config]
 
   snark_worker_replicas   = var.snark_worker_replicas
   snark_worker_fee        = var.snark_worker_fee
@@ -57,6 +45,8 @@ module "kubernetes_testnet" {
       run_with_user_agent    = false
       run_with_bots          = false
       enable_peer_exchange   = true
+      enableArchive          = false
+      archiveAddress         = null
     }
   ]
 }