+++ src/lib/rosetta_lib/errors.ml
@@ -36,7 +36,8 @@ module Variant = struct
     | `No_options_provided
     | `Exception of string
     | `Signature_invalid
-    | `Memo_invalid ]
+    | `Memo_invalid
+    | `Graphql_uri_not_set ]
   [@@deriving yojson, show, eq, to_enum, to_representatives]
 end
 
@@ -104,6 +105,8 @@ end = struct
         "Invalid signature"
     | `Memo_invalid ->
         "Invalid memo"
+    | `Graphql_uri_not_set ->
+        "No GraphQL URI set"
 
   let context = function
     | `Sql msg ->
@@ -166,6 +169,8 @@ end = struct
         None
     | `Memo_invalid ->
         None
+    | `Graphql_uri_not_set ->
+        None
 
   let retriable = function
     | `Sql _ ->
@@ -203,6 +208,8 @@ end = struct
     | `Signature_invalid ->
         false
     | `Memo_invalid ->
+        false
+    | `Graphql_uri_not_set ->
         false
 
   (* Unlike message above, description can be updated whenever we see fit *)
@@ -243,6 +250,9 @@ end = struct
         "Your request has an invalid memo."
     | `No_options_provided ->
         "Your request is missing options."
+    | `Graphql_uri_not_set ->
+        "This Rosetta instance is running without a GraphQL URI set but this \
+         request requires one."
     | `Exception _ ->
         "We encountered an internal exception while processing your request. \
          (That means you found a bug!)"