+++ helm/block-producer/CHANGELOG.md
@@ -15,3 +15,10 @@
 ------
 - add peers list URL chart value
 
+1.0.0:
+------
+- coda -> mina
+
+1.0.1:
+------
+- enable daemon blockchain archival capability