+++ helm/snark-worker/CHANGELOG.md
@@ -11,3 +11,11 @@
 ------
 - add peers list URL chart value
 
+1.0.0:
+------
+- coda -> mina
+
+1.0.1:
+------
+- enable daemon blockchain archival capability
+